[fix] addressed comments: removed unused imports, reformatted import path directories, and removed local styles from global styles file
